+++ MelonLoader/Melons/FileTypes/DLL.cs
@@ -13,9 +13,9 @@ namespace MelonLoader.MelonFileTypes
                 ? MelonLaunchOptions.Core.LoadMode_Plugins
                 : MelonLaunchOptions.Core.LoadMode_Mods;
             string[] filearr = Directory.GetFiles(folderpath).Where(x => 
-                Path.GetExtension(x).Equals(".dll")
-                && ((loadMode == MelonLaunchOptions.Core.LoadModeEnum.DEV) ? x.EndsWith(".dev.dll")
-                : ((loadMode == MelonLaunchOptions.Core.LoadModeEnum.NORMAL) ? !x.EndsWith(".dev.dll")
+                Path.GetExtension(x).ToLowerInvariant().Equals(".dll")
+                && ((loadMode == MelonLaunchOptions.Core.LoadModeEnum.DEV) ? x.ToLowerInvariant().EndsWith(".dev.dll")
+                : ((loadMode == MelonLaunchOptions.Core.LoadModeEnum.NORMAL) ? !x.ToLowerInvariant().EndsWith(".dev.dll")
                 : true))
             ).ToArray();
             if (filearr.Length <= 0)