+++ MelonLoader/Extensions/Il2CppAssemblyGenerator.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.IO;
 using System.Reflection;
+using System.Runtime.CompilerServices;
 
 namespace MelonLoader
 {
@@ -18,7 +19,9 @@ namespace MelonLoader
 
             if (RunMethod != null)
             {
+                DisableCloseButton();
                 int returnval = (int)RunMethod.Invoke(null, new object[0]);
+                EnableCloseButton();
                 Fixes.ApplicationBase.Run(AppDomain.CurrentDomain);
                 return (returnval == 0);
             }
@@ -69,5 +72,10 @@ namespace MelonLoader
             }
             catch (Exception ex) { MelonLogger.ThrowInternalFailure($"Il2CppAssemblyGenerator Exception: {ex}"); }
         }
+
+        [MethodImpl(MethodImplOptions.InternalCall)]
+        private extern static void EnableCloseButton();
+        [MethodImpl(MethodImplOptions.InternalCall)]
+        private extern static void DisableCloseButton();
     }
 }