+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 1. Fixed Issue with Compatibility Layer system throwing NRE about Default Constructors.
 2. Added MelonCompatibilityLayer.WrapperData Extensions.
 3. Fixed Case Issue with MelonHandler not handling File Extensions properly.
+4. Fixed Case Issue with MelonHandler not handling File Extensions inside ZIPs properly.
 
 ---